Animation Changes
I added a Animation Tree since I wanted to simplify my code and make it easier to add new animations later on. I made a animated state machine for the player to play animation correctly when throwing and running. This is a faster and overall better solution for player animations.